requirements.txt
# requirements.txt
google-adk
fastapi
uvicorn
requests
click
pydantic
litellm
python-docx 